+++ cdxev/auxiliary/identity.py
@@ -2,10 +2,16 @@
 
 import functools
 import json
+import logging
 import re
 import typing as t
-from dataclasses import dataclass
+from dataclasses import dataclass, fields
 from enum import Enum
+
+import univers.version_range
+import univers.versions
+
+from cdxev.log import LogMessage
 
 
 @functools.total_ordering
@@ -201,6 +207,164 @@ class RegexUpdateIdentity:
 
     def __str__(self) -> str:
         return f"{self.field}[regex:{self.expression}]"
+
+
+@dataclass(init=True, frozen=True)
+class CoordinatesWithVersionRange(Coordinates):
+    """
+    This class inherits from cdxev.auxiliary.identity.Coordinates
+    and extends it to be able to handle version ranges.
+    It achieves this by introducing the attribute 'version_range'
+    of type univers.version_range.VersionRange.
+
+    This class's comparator is compatible with Coordinates objects.
+    If the 'name' and 'group' match,
+    it will check if the provided version of the coordinates object
+    is contained in this class instance's 'version_range'
+    and return True or False depending on the result.
+    """
+
+    version_range: univers.version_range.VersionRange
+
+    def __eq__(self, other: object) -> bool:
+        if t.TYPE_CHECKING:
+            # univers guarantees that version_class is not None
+            assert self.version_range.version_class is not None  # nosec - only for type checker
+
+        if isinstance(other, CoordinatesWithVersionRange):
+            for field_ in fields(self):
+                if getattr(self, field_.name) != getattr(other, field_.name):
+                    return False
+            return True
+
+        if isinstance(other, Coordinates):
+            for field_ in fields(other):
+                if field_.name != "version" and getattr(self, field_.name) != getattr(
+                    other, field_.name
+                ):
+                    return False
+
+            if other.version is not None:
+                try:
+                    if self.version_range.version_class(other.version) in self.version_range:
+                        return True
+                except univers.versions.InvalidVersion:
+                    possible_versions = []
+                    for version_type in univers.versions.AVAILABLE_VERSIONS:
+                        try:
+                            if version_type.is_valid(other.version):
+                                possible_versions.append(str(version_type.__name__))
+                        except univers.versions.nuget.InvalidNuGetVersion:
+                            pass
+                    version_is_of = " which is valid under the schemas: "
+
+                    if not possible_versions:
+                        version_is_of = "which versioning schema is not supported"
+                    else:
+                        for version in possible_versions:
+                            version_is_of += version + ", "
+                        version_is_of = version_is_of[:-2]
+                    logger = logging.getLogger(__name__)
+                    logger.warning(
+                        LogMessage(
+                            "Incompatible version ranges",
+                            f"The component {other} matches the target {self}"
+                            f" in the name and group keys but has a different versioning"
+                            f" schema. The target has versioning schema"
+                            f' "{self.version_range.version_class.__name__}"'
+                            f' this is incompatible with the version "{other.version}"'
+                            + version_is_of,
+                        )
+                    )
+                    return False
+
+        return False
+
+    def __str__(self) -> str:
+        group_str = (f"{self.group}/") if self.group is not None else ""
+        version_str = (f"@{self.version_range}") if self.version_range is not None else ""
+        return group_str + self.name + version_str
+
+
+@dataclass(frozen=True)
+class CoordinatesRegexIdentity:
+    """
+    Matches components by a regex on *name*, optionally narrowed by
+    *group* (literal string or regex), *version* (exact), or *version-range*.
+    """
+
+    name_expression: str
+    name_pattern: re.Pattern[str]
+    group_expression: t.Optional[str]
+    group_pattern: t.Optional[re.Pattern[str]]
+    version: t.Optional[str]
+    version_range: t.Optional[univers.version_range.VersionRange]
+
+    @classmethod
+    def create(
+        cls,
+        name_expression: str,
+        group_expression: t.Optional[str] = None,
+        group_is_regex: bool = False,
+        version: t.Optional[str] = None,
+        version_range: t.Optional[univers.version_range.VersionRange] = None,
+    ) -> "CoordinatesRegexIdentity":
+        name_pattern = re.compile(f"^(?:{name_expression})$")
+        if group_expression is not None:
+            grp_str = (
+                f"^(?:{group_expression})$"
+                if group_is_regex
+                else f"^{re.escape(group_expression)}$"
+            )
+            group_pat: t.Optional[re.Pattern[str]] = re.compile(grp_str)
+        else:
+            group_pat = None
+        return cls(
+            name_expression=name_expression,
+            name_pattern=name_pattern,
+            group_expression=group_expression,
+            group_pattern=group_pat,
+            version=version,
+            version_range=version_range,
+        )
+
+    def matches(self, component: dict) -> bool:
+        name = component.get("name")
+        if not isinstance(name, str) or not self.name_pattern.fullmatch(name):
+            return False
+
+        if self.group_pattern is not None:
+            comp_group = component.get("group")
+            if not isinstance(comp_group, str) or not self.group_pattern.fullmatch(comp_group):
+                return False
+
+        if self.version is not None:
+            if component.get("version") != self.version:
+                return False
+
+        if self.version_range is not None:
+            if t.TYPE_CHECKING:
+                assert self.version_range.version_class is not None  # nosec
+            comp_version = component.get("version")
+            if comp_version is None:
+                return False
+            try:
+                if self.version_range.version_class(comp_version) not in self.version_range:
+                    return False
+            except univers.versions.InvalidVersion:
+                return False
+
+        return True
+
+    def __str__(self) -> str:
+        group_str = f"/{self.group_expression}" if self.group_expression is not None else ""
+        if self.version is not None:
+            ver_str = f"@{self.version}"
+        elif self.version_range is not None:
+            ver_str = f"@{self.version_range}"
+        else:
+            ver_str = ""
+        return f"name[regex:{self.name_expression}]{group_str}{ver_str}"
 
 
 @dataclass(frozen=True, init=True)
